Support round-tripping line numbers on lists
This enables macros to reuse input params and line numbers to support
use cases like ‘defn’ where the macro defines a fn and reuses the
arguments as the fn body.

See the 'example' procedure in core.lsp for a working example of line
numbers surviving macro expansion, resulting in a stack trace with
accurate line numbers. No automated tests written for this behavior yet.

May expand metadata support to all allocated value types in the future.

- Compiler embeds line numbers into list constants
- VM models metadata properties on Cons objects as a list of key-value
  pairs
- VM hydrates line-numbers as metadata properties on lists
- VM prn’s eval result, embeds line numbers found  in list metadata