fixed gc issues, removed longjmp usage
Using `longjmp` makes it difficult to correctly clean up resources, including the gc-roots programmatically pushed and popped by c code inside the VM. In this case, the first exception left a dangling root on the first frame which pointed to a stack value. The first exception handler also fails, while pushing another dangling root into the second frame (points to the same places as the first dangling root).

When GC was running, it would relocate the first root it found, thereby indirectly relocating the second. And when it encountered the second, it attempted to relocate that too, even though it was already moved to the current heap. This broke things.

The solution was to
1. put in a check so only things from the old heap can be relocated to the new one

2. stop using longjmp, and go back to using return values as indicators of raising an exception

This seems to work fine. I am appropriately chastised.