rework analyzer to reduce 'constant' usage
The analyzer and compiler assume that whatever gets quoted should be literally loadable into the VM as-is. This means that the VM has to have native support for whatever gets quoted.

This change reduces these assumptions within the analyzer, paving the way for reducing the mandadory surface area needed in the VM.

This doesn't address the macro-expansion mechanism, which seems to rely on this most heavily.

For clarity, an initial goal here is to remove the need for hash-map constant support from the VM hydration code.